Move Concrete types to Concrete namespace
