+++ src/Service/Lexicon/V1/ComponentGenerator/Field/BooleanComponentGenerator.php
@@ -7,7 +7,7 @@ use Blugen\Service\Lexicon\ArraySerialization\ArraySerializationContext;
 use Blugen\Service\Lexicon\ArraySerialization\ArraySerializationContributor;
 use Blugen\Service\Lexicon\GeneratorInterface;
 use Blugen\Service\Lexicon\V1\Property;
-use Blugen\Service\Lexicon\V1\Schema\Field\BooleanSchema;
+use Blugen\Service\Lexicon\V1\Schema\Concrete\BooleanSchema;
 use Nette\PhpGenerator\ClassType;
 
 class BooleanComponentGenerator implements GeneratorInterface, ArraySerializationContributor
