+++ src/Service/Lexicon/V1/ComponentGenerator/Field/BytesComponentGenerator.php
@@ -7,7 +7,7 @@ use Blugen\Service\Lexicon\ArraySerialization\ArraySerializationContext;
 use Blugen\Service\Lexicon\ArraySerialization\ArraySerializationContributor;
 use Blugen\Service\Lexicon\GeneratorInterface;
 use Blugen\Service\Lexicon\V1\Property;
-use Blugen\Service\Lexicon\V1\Schema\Field\BytesSchema;
+use Blugen\Service\Lexicon\V1\Schema\Concrete\BytesSchema;
 use Nette\PhpGenerator\ClassType;
 
 class BytesComponentGenerator implements GeneratorInterface, ArraySerializationContributor
