+++ src/Service/Lexicon/V1/ComponentGenerator/Field/StringComponentGenerator.php
@@ -7,7 +7,7 @@ use Blugen\Service\Lexicon\ArraySerialization\ArraySerializationContext;
 use Blugen\Service\Lexicon\ArraySerialization\ArraySerializationContributor;
 use Blugen\Service\Lexicon\GeneratorInterface;
 use Blugen\Service\Lexicon\V1\Property;
-use Blugen\Service\Lexicon\V1\Schema\Field\StringSchema;
+use Blugen\Service\Lexicon\V1\Schema\Concrete\StringSchema;
 use Nette\PhpGenerator\ClassType;
 
 class StringComponentGenerator implements GeneratorInterface, ArraySerializationContributor
