+++ src/Service/Syntax/Factory/SchemaFactory.php
@@ -4,16 +4,16 @@ namespace Blugen\Service\Syntax\Factory;
 
 use Blugen\Service\Lexicon\SchemaInterface;
 use Blugen\Service\Lexicon\V1\Schema;
+use Blugen\Service\Lexicon\V1\Schema\Concrete\BlobSchema;
+use Blugen\Service\Lexicon\V1\Schema\Concrete\BooleanSchema;
+use Blugen\Service\Lexicon\V1\Schema\Concrete\BytesSchema;
+use Blugen\Service\Lexicon\V1\Schema\Concrete\CidLinkSchema;
+use Blugen\Service\Lexicon\V1\Schema\Concrete\IntegerSchema;
+use Blugen\Service\Lexicon\V1\Schema\Concrete\NullSchema;
+use Blugen\Service\Lexicon\V1\Schema\Concrete\StringSchema;
 use Blugen\Service\Lexicon\V1\Schema\Field\ArraySchema;
-use Blugen\Service\Lexicon\V1\Schema\Field\BlobSchema;
-use Blugen\Service\Lexicon\V1\Schema\Field\BooleanSchema;
-use Blugen\Service\Lexicon\V1\Schema\Field\BytesSchema;
-use Blugen\Service\Lexicon\V1\Schema\Field\CidLinkSchema;
-use Blugen\Service\Lexicon\V1\Schema\Field\IntegerSchema;
-use Blugen\Service\Lexicon\V1\Schema\Field\NullSchema;
 use Blugen\Service\Lexicon\V1\Schema\Field\ObjectSchema;
 use Blugen\Service\Lexicon\V1\Schema\Field\RefSchema;
-use Blugen\Service\Lexicon\V1\Schema\Field\StringSchema;
 use Blugen\Service\Lexicon\V1\Schema\Field\TokenSchema;
 use Blugen\Service\Lexicon\V1\Schema\Field\UnionSchema;
 use Blugen\Service\Lexicon\V1\Schema\Field\UnknownSchema;
