+++ tests/Unit/Service/Lexicon/V1/TypeSpecificSchema/Field/BooleanSchemaTest.php
@@ -3,7 +3,7 @@
 namespace Blugen\Tests\Unit\Service\Lexicon\V1\TypeSpecificSchema\Field;
 
 use Blugen\Service\Lexicon\V1\Schema;
-use Blugen\Service\Lexicon\V1\Schema\Field\BooleanSchema;
+use Blugen\Service\Lexicon\V1\Schema\Concrete\BooleanSchema;
 use PHPUnit\Framework\TestCase;
 
 class BooleanSchemaTest extends TestCase
